refactoring and changing styles
